test: add test for uploadFileCalls
