+++ packages/amplify-category-hosting/__tests__/lib/S3AndCloudFront/helpers/file-uploader.test.js
@@ -2,15 +2,17 @@ jest.mock('mime-types');
 jest.mock('../../../../lib/S3AndCloudFront/helpers/file-scanner', () => {
   return {
     scan: jest.fn(() => {
-      return ['filePath1', 'filePath2'];
+      return ['assets/index-4e7fb3f6.js', 'index.html', 'assets/mapbox-gl-d152561b.js', 'assets/index-d41dc7ea.css', 'vite.svg'];
     }),
   };
 });
+jest.mock('../../../../lib/S3AndCloudFront/helpers/upload-file', () => {return {uploadFile: jest.fn()}});
 
 const fs = require('fs-extra');
 const mime = require('mime-types');
 
 const fileScanner = require('../../../../lib/S3AndCloudFront/helpers/file-scanner');
+const { uploadFile } = require('../../../../lib/S3AndCloudFront/helpers/upload-file');
 
 const mockTemplate = require('../../../../__mocks__/mockTemplate');
 const mockParameters = require('../../../../__mocks__/mockParameters');
@@ -80,19 +82,27 @@ describe('cloudfront-manager', () => {
     },
   };
 
-  beforeAll(() => {
+  beforeAll(async () => {
     fs.createReadStream = jest.fn(() => {
       return {};
     });
     mime.lookup = jest.fn(() => {
       return 'text/plain';
     });
+    await fileUploader.run(mockContext, 'mockDistributionFolder');
   });
 
   beforeEach(() => {});
 
-  test('run', async () => {
-    await fileUploader.run(mockContext, 'mockDistributionFolder');
+  test('file scanner should be called', async () => {
     expect(fileScanner.scan).toBeCalled();
   });
+
+  test('uploadFileCalls should have the index.html at the end', () => { 
+    expect(uploadFile).toHaveBeenCalled();
+    const uploadFileCalls = uploadFile.mock.calls;
+
+    expect(uploadFileCalls.length).toBe(5);
+    expect(uploadFileCalls[4][3]).toBe('index.html');
+  })
 });
